fixed int uint comparison warning
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@8236 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-gui/src/Workspace.cpp
@@ -283,7 +283,7 @@ void Workspace::updatePanels() {
   if (_currentPanelIndex<0)
     _currentPanelIndex=0;
 
-  if (_currentPanelIndex > _panels.size()-currentSlotsCount())
+  if ((uint)_currentPanelIndex > _panels.size()-currentSlotsCount())
     _currentPanelIndex = _panels.size()-currentSlotsCount();
 
 //   Fill up slots according to the current index until there is no panel to show
